+++ doc/release-notes.md
@@ -9,3 +9,5 @@ This release includes the following features and fixes:
    several months and has been removed completely.
  - The `getstakingreward` RPC now returns the `proofid` of the staking reward
    winner in addition to the payout script.
+ - A new `setflakyproof` RPC instructs the node to also accept an alternative
+   staking reward winner when the flaky proof would have been selected.